(feat: 펫 이미지 가져와서 네비게이션 바에 표시)
