+++ src/layouts/dashboard/nav/index.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 // @mui
 import { styled, alpha } from '@mui/material/styles';
@@ -14,6 +14,8 @@ import Scrollbar from '../../../components/scrollbar';
 import NavSection from '../../../components/nav-section';
 //
 import navConfig from './config';
+import Cookies from 'js-cookie';
+import api from '../../../lib/api';
 
 // ----------------------------------------------------------------------
 
@@ -36,6 +38,7 @@ Nav.propTypes = {
 
 export default function Nav({ openNav, onCloseNav }) {
   const { pathname } = useLocation();
+  const [petImg, setPetImg] = useState();
 
   const isDesktop = useResponsive('up', 'lg');
 
@@ -45,6 +48,32 @@ export default function Nav({ openNav, onCloseNav }) {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [pathname]);
+
+  useEffect(() => {
+    handlePetUrl()
+  }, [petImg])
+
+    //헤더 아바타에 들어갈 펫 이미지 
+    const handlePetUrl = () => {
+      if (Cookies.get("key")) {
+        api.get("pet/petinfo")
+          .then((res) => {
+            console.log("res.data " + res.data.data.petUrl)
+            setPetImg(res.data.data.petUrl);
+          }).catch((error) => {
+            api.post("member/reissue")
+              .then((res) => {
+                console.log("accesstoken" + res.data.data);
+                Cookies.set("key", res.data.data);
+                console.log('토큰 재발급 성공');
+              })
+              .catch((err) => {
+                console.log(err.message)
+              })
+            console.log(error.message)
+          })
+      }
+    }
 
   const renderContent = (
     <Scrollbar
@@ -60,7 +89,7 @@ export default function Nav({ openNav, onCloseNav }) {
       <Box sx={{ mb: 5, mx: 2.5 }}>
         <Link underline="none">
           <StyledAccount>
-            <Avatar src={account.photoURL} alt="photoURL" />
+            <Avatar src={petImg} alt="photoURL" />
 
             <Box sx={{ ml: 2 }}>
               <Typography variant="subtitle2" sx={{ color: 'text.primary' }}>
