+++ src/routes.js
@@ -3,18 +3,20 @@ import { Navigate, useRoutes } from 'react-router-dom';
 import DashboardLayout from './layouts/dashboard';
 import SimpleLayout from './layouts/simple';
 //
+import LoginPage from './pages/LoginPage';
 import TodoPage from './pages/Todo';
 import BoardPage from './pages/BoardListPage';
 //import BoardFormPage from 'pages/BoardFormPage';
-import BoardCreateForm from 'pages/BoardCreatePage'
-import BoardModifyForm from 'pages/BoardModifyPage'
-import BoardDetail from 'pages/BoardDetailPage'
+import BoardCreateForm from './pages/BoardCreatePage'
+import BoardModifyForm from './pages/BoardModifyPage'
+import BoardDetail from './pages/BoardDetailPage'
 import BlogPage from './pages/BlogPage';
 import UserPage from './pages/UserPage';
-import LoginPage from './pages/LoginPage';
+//import LoginPage from './pages/LoginPage';
 import Page404 from './pages/Page404';
 import ProductsPage from './pages/ProductsPage';
 import DashboardAppPage from './pages/DashboardAppPage';
+import PetFormPage from './pages/PetFormPage';
 
 // ----------------------------------------------------------------------
 
@@ -41,6 +43,14 @@ export default function Router() {
         { element: <Navigate to="/dashboard/app" />, index: true },
         { path: '404', element: <Page404 /> },
         { path: '*', element: <Navigate to="/404" /> },
+      ],
+    },
+    {
+      path: '/petform',
+      element: <DashboardLayout />,
+      children: [
+        { element: <PetFormPage />, index: true },
+        { path: 'petform', element: <PetFormPage /> },
       ],
     },
     {
